refactor(entities, widgets): changed of import entities to widget
