+++ src/entities/blog/ui/post-main-card.tsx
