+++ src/entities/blog/ui/post-mobile-card.tsx
